+++ app/src/main/java/dev/arkbuilders/rate/data/db/entity/RoomQuickPair.kt
@@ -4,6 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 import dev.arkbuilders.rate.domain.model.Amount
 import dev.arkbuilders.rate.domain.model.CurrencyCode
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 
 @Entity
@@ -11,7 +12,7 @@ data class RoomQuickPair(
     @PrimaryKey(autoGenerate = true)
     val id: Long,
     val from: CurrencyCode,
-    val amount: Double,
+    val amount: BigDecimal,
     val to: List<Amount>,
     val calculatedDate: OffsetDateTime,
     val pinnedDate: OffsetDateTime?,
